api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-cloud-bigquery==3.13.0
google-cloud-aiplatform==1.38.1
google-cloud-vertexai==1.38.1
neo4j==5.14.1
pydantic==2.5.0
